+++ lms/templates/lti_form.html
@@ -12,7 +12,7 @@ from openedx.core.djangolib.js_utils import js_escaped_string
     </head>
     <body>
     ## This form will be hidden.
-    ## LTI module JavaScript will trigger a "submit" on the form, and the
+    ## LTI block JavaScript will trigger a "submit" on the form, and the
     ## result will be rendered instead.
         <form
             id="lti-${element_id}"
