+++ setup.py
@@ -23,7 +23,7 @@ XBLOCKS = [
     "library = xmodule.library_root_xblock:LibraryRoot",
     "library_content = xmodule.library_content_block:LibraryContentBlock",
     "library_sourced = xmodule.library_sourced_block:LibrarySourcedBlock",
-    "lti = xmodule.lti_module:LTIBlock",
+    "lti = xmodule.lti_block:LTIBlock",
     "poll_question = xmodule.poll_module:PollBlock",
     "problem = xmodule.capa_block:ProblemBlock",
     "randomize = xmodule.randomize_module:RandomizeBlock",
