+++ xmodule/static_content.py
@@ -24,7 +24,7 @@ from xmodule.capa_block import ProblemBlock
 from xmodule.conditional_block import ConditionalBlock
 from xmodule.html_block import AboutBlock, CourseInfoBlock, HtmlBlock, StaticTabBlock
 from xmodule.library_content_block import LibraryContentBlock
-from xmodule.lti_module import LTIBlock
+from xmodule.lti_block import LTIBlock
 from xmodule.poll_module import PollBlock
 from xmodule.seq_module import SequenceBlock
 from xmodule.split_test_module import SplitTestBlock
